moved page class into pages package
